Move tests from EvalErrorTests to the functions they actually test
